social-media-automation/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
PyMuPDF==1.23.0
schedule==1.2.0
pyyaml==6.0
